*특좀 (자키)추가 *적 정보데이터 파일 추가 / 형식수정 *네비메쉬 변경

+++ Assets/Scripts/TestEnemySpawner.cs
@@ -81,10 +81,10 @@ public class TestEnemySpawner : MonoBehaviour
 
     private void AddToPool(TestEnemy enemy)
     {
-        if (!Pool.TryGetValue(enemy.MyEnemyType, out var list))
+        if (!Pool.TryGetValue((int)enemy.MyEnemyType, out var list))
         {
             list = new List<TestEnemy>();
-            Pool.Add(enemy.MyEnemyType, list);
+            Pool.Add((int)enemy.MyEnemyType, list);
         }
         list.Add(enemy);
     }
